FIX: I (hopefully) fixed issue #154 regarding a blocking exception when refreshing the event log (also modified the DataView binding as it was assigned by code instead of XAML). IMP: Removed ResolvedHost from the CurrentConn class, as we already have a TargetHostName property on the parent class with the same goal. CurrentConn will be simplified further (and eventually renamed). MISC: minor refactoring.

+++ Common/Processes/ProcessHelper.cs
@@ -397,7 +397,7 @@ public static partial class ProcessHelper
     public static string GetLocalUserOwner(uint pid)
     {
         //Based on: https://bytes.com/topic/c-sharp/answers/225065-how-call-win32-native-api-gettokeninformation-using-c
-        IntPtr hProcess = NativeMethods.OpenProcess(NativeMethods.ProcessAccessFlags.QueryInformation, false, (uint)pid);
+        IntPtr hProcess = NativeMethods.OpenProcess(NativeMethods.ProcessAccessFlags.QueryInformation, false, pid);
         if (hProcess == IntPtr.Zero)
         {
             LogHelper.Warning($"Unable to retrieve process local user owner: process pid={pid} cannot be found!");
